+++ _/tests/internal/sub_one/mod.rs
@@ -1,3 +1,0 @@
-use import_modules::import_pub_modules;
-
-import_pub_modules!("tests/internal/sub_one/", "^((?!mod.rs).)*$");