+++ _/tests/priv_modules/a.rs
@@ -1,3 +1,0 @@
-pub fn test() -> &'static str {
-    "The modules are imported!"
-}